Demo printing after map update
Update the identify point plugin print to with an asynchronous action
(adding a layer to the map). This requires waiting for the map to update
before proceeding with printing.

The timeout is extended to give the map object time to realize it needs
to update. Proceeding too quickly results in the layer being added after
printing.

This implementation is largely based off of https://github.com/CoastalResilienceNetwork/future_habitat_v2

Refs #1062