remove min-height from tinner class
Class tinner is a wrapper div for the Tiny box popups. The framework
adds a min-height of 380px to this wrapper. If you need a popup with a
height less-than 380px it adds unwanted white-space at the bottom of the
popup. Removing or significantly lowering the min-height would give more
control to the height of the popup windows.