Demo print preview in plugin print workflow
Add a print preview map to the identify point plugin. This serves as
a demo to plugin developers of how to add a print preview map to the
plugin print workflow.

Also, make some adjustments to the optional print images and map size
for the plugin to compensate for the new sizing units.

Refs #1103