Ensure printed map view matches app map view
For the plugin print process, the printed map center and extent should
match that of the map in the app. TNC deemed this as the best UX.

Note that this will happen regardless if there is a print preview map or
not. If there is a print preview map, the printed map will match the
center and extent of the that map.

Refs #1113