Print plugin demo clean-up
Ensures the document is only one page when printed in major browsers by
by making the document take up 90% of the page, instead of 100%, which
leaves room for margins.

Also removes some rules that were no longer applicable and makes sure the
north arrow is not covered by the legend.

Refs #1010