Template out plugin initialization and assets
Follow-up to #1127

Migrate plugin template variable creation to python scripts, and add
plugin initialization and CSS asset includes to index template. The
function merge_plugin_config_data is basically a port of
MergePluginConfigurationData from the C# code.

Also, the context variable for the template had to be reworked a bit
so that the region.json variables could be referred to individually
and as a whole.

Refs #1099