Fix path to font files
For some reason, the hash on the path to the font files causes the
request to 404 when using a static server. A similar error has been
reported to Fontello, and their suggested workaround was to remove the
hash from the static CSS file.

https://github.com/fontello/fontello/issues/455

Note that our Fontello distribution is unlikely to change, so I don't
forsee reimplementing this for every Fontello update as a big issue.

Refs #1108