Fix single plugin mode template conditionals
Test against the "defined" value, which is provided by Jinja, to see if
the single plugin mode configuration block exists in the region.json.
This is suggested by Jinja for testing of existence of a variable.

Also, some conditionals needed to be reversed so that the normal
framework mode had all of its components included.