Use json file extension for translation files
The previous implementation used a dynamic endpoint to serve the
translation file, which didn't have a file extension. When we initially
migrated the translation backend to the static build, we left the file
without an extension. This worked for most installations, but proved to
be problematic when serving the app from IIS, which uses a whitelist for
serving file extensions. Adding the json file extension also makes more
practical sense than using no extension since the file is indeed JSON.

Refs #1163