Clean-up mobile non-single app mode
Most of the improvements made to mobile single app mode should apply to
normal mobile mode. However, there are a few rules that caused problems
in normal mode, as well as some code and elements that do not need to
exist or have an effect in normal mode. The offending rules, code, and
elements inclusion are all made conditional on single app mode being
active.

Refs #1143