add constant Google Analytics ID
main.py passes this doc a var that tells it if the build is intended for prod.  Google Analytics will only be activated if the build.py script is told to puch the build to prod. This prevents dev and sandbox builds from being logged. The GA ID is now a constant and the next step will be to remove that property from region.json and check if removing the property breaks the build.