hardcode headerLogoPath
This means we no longer need to specify headerLogoPath in the print object of each region's region.json file. We do need to leave the print object in region.json for now, as it is a required property. We can simply add it as "print":{} (empty object). I will look into changing print to an optional property.